Prevented a problem when unloading players, and when a player is null. The condition that was causing a null player was that the player was unloaded when the player left the server at the same time when the server was shut down.  Basically a race condition with two parallel async tasks trying to shut down the player cache object, each focusing on a different aspect (player vs. server).

+++ prison-core/src/main/java/tech/mcprison/prison/cache/PlayerCacheFiles.java
@@ -90,36 +90,39 @@ public class PlayerCacheFiles
 	 */
 	public void toJsonFile( PlayerCachePlayerData player) {
 		
-		File playerFile = player.getPlayerFile();
-		File outTemp = createTempFile( playerFile );
-		boolean success = false;
-		
-		try (
-				FileWriter fw = new FileWriter( outTemp );
-				){
-			getGson().toJson( player, fw );
+		if ( player != null ) {
 			
-			success = true;
-		}
-		catch ( JsonIOException | IOException e ) {
-			e.printStackTrace();
-		}
-		
-		if ( success && ( !playerFile.exists() || playerFile.delete()) ) {
-			outTemp.renameTo( playerFile );
-		}
-		else {
-		
-			boolean removed = false;
-			if ( outTemp.exists() ) {
-				removed = outTemp.delete();
+			File playerFile = player.getPlayerFile();
+			File outTemp = createTempFile( playerFile );
+			boolean success = false;
+			
+			try (
+					FileWriter fw = new FileWriter( outTemp );
+					){
+				getGson().toJson( player, fw );
+				
+				success = true;
 			}
-
-			String message = String.format( 
-					"Unable to rename PlayerCache temp file. It was %sremoved: %s", 
-					(removed ? "" : "not "), outTemp.getAbsolutePath() );
+			catch ( JsonIOException | IOException e ) {
+				e.printStackTrace();
+			}
 			
-			Output.get().logWarn( message );
+			if ( success && ( !playerFile.exists() || playerFile.delete()) ) {
+				outTemp.renameTo( playerFile );
+			}
+			else {
+				
+				boolean removed = false;
+				if ( outTemp.exists() ) {
+					removed = outTemp.delete();
+				}
+				
+				String message = String.format( 
+						"Unable to rename PlayerCache temp file. It was %sremoved: %s", 
+						(removed ? "" : "not "), outTemp.getAbsolutePath() );
+				
+				Output.get().logWarn( message );
+			}
 		}
 	}
